requirements.txt
Django==3.2.8
djangorestframework==3.12.4
django-cors-headers==3.7.0
django-channels==3.0.4
channels-redis==3.3.1
celery==5.1.2
openai==0.27.0
psycopg2-binary==2.9.1
gunicorn==20.1.0
django-redis==5.0.0
elasticsearch==7.15.1
python-dotenv==0.19.1
django-ratelimit==3.0.1
django-extensions==3.1.3
django-elasticsearch-dsl==7.2.1
kibana==1.0.0
numpy==1.21.2
pandas==1.3.3
scikit-learn==0.24.2
scipy==1.7.1
tensorflow==2.6.0
keras==2.6.0
